Refactor project structure and migrate to Next.js for development environment
- Removed outdated files including the Express API server and Vite configuration, transitioning to a Next.js architecture.
- Introduced new components and actions for predictive text functionality, including a DevPlayground component for testing.
- Updated package.json to reflect Next.js scripts and dependencies, enhancing the development environment.
- Streamlined TypeScript configuration and added necessary files for Next.js compatibility.
- Enhanced styling and structure for improved organization and usability in the development playground.
