--- dev/index.html
@@ -1,111 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <link rel="icon" type="image/svg+xml" href="/vite.svg" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>PredictiveTextarea - Development</title>
-  <style>
-    body {
-      font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
-      max-width: 1200px;
-      margin: 0 auto;
-      padding: 2rem;
-      background-color: #f9fafb;
-      color: #111827;
-    }
-
-    h1 {
-      font-size: 1.875rem;
-      font-weight: bold;
-      margin-bottom: 1.5rem;
-    }
-
-    .grid {
-      display: grid;
-      grid-template-columns: 1fr 1fr;
-      gap: 1.5rem;
-    }
-
-    .card {
-      background-color: #fff;
-      border-radius: 0.5rem;
-      padding: 1.5rem;
-      box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-    }
-
-    .space-y-4>*+* {
-      margin-top: 1rem;
-    }
-
-    .form-group label {
-      display: block;
-      font-size: 0.875rem;
-      font-weight: 500;
-      margin-bottom: 0.5rem;
-    }
-
-    input,
-    select {
-      width: 100%;
-      height: 2.25rem;
-      padding: 0 0.75rem;
-      border-radius: 0.375rem;
-      border: 1px solid #d1d5db;
-      background-color: #fff;
-      font-size: 0.875rem;
-    }
-
-    input:focus,
-    select:focus {
-      outline: none;
-      border-color: #3b82f6;
-      box-shadow: 0 0 0 2px rgba(59, 130, 246, 0.3);
-    }
-
-    .checkbox-group {
-      display: flex;
-      align-items: center;
-      gap: 0.5rem;
-    }
-
-    .checkbox-group input {
-      width: auto;
-      height: auto;
-    }
-
-    .custom-textarea {
-      min-height: 200px;
-      width: 100%;
-      border: 1px solid #d1d5db;
-      border-radius: 0.375rem;
-      background-color: #fff;
-      padding: 0.75rem;
-      font-size: 0.875rem;
-    }
-
-    .prediction-indigo {
-      color: #6366f1;
-      opacity: 0.6;
-    }
-
-    .prediction-subtle {
-      color: #9ca3af;
-      opacity: 0.7;
-    }
-
-    .prediction-fancy {
-      color: #3b82f6;
-      font-style: italic;
-      opacity: 0.8;
-    }
-  </style>
-</head>
-
-<body>
-  <div id="root"></div>
-  <script type="module" src="/src/main.tsx"></script>
-</body>
-
-</html>
