+++ dev/package.json
@@ -1,12 +1,12 @@
 {
   "name": "dev",
-  "type": "module",
   "private": true,
   "scripts": {
-    "dev": "vite",
+    "dev": "next dev",
     "dev:api": "node api-server.js",
-    "build": "tsc && vite build",
-    "preview": "vite preview"
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
   },
   "dependencies": {
     "@radix-ui/react-slot": "^1.0.0",
@@ -17,6 +17,7 @@
     "cors": "^2.8.5",
     "express": "^5.1.0",
     "lucide-react": "^0.344.0",
+    "next": "^14.1.0",
     "openai": "^4.95.1",
     "prism-react-renderer": "^2.4.1",
     "react": "^18.2.0",
@@ -29,12 +30,12 @@
     "@types/node": "^20.11.0",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
-    "@vitejs/plugin-react": "^4.2.0",
     "autoprefixer": "^10.4.0",
     "concurrently": "^9.1.2",
+    "eslint": "^8.0.0",
+    "eslint-config-next": "^14.1.0",
     "postcss": "^8.4.0",
-    "typescript": "^5.0.0",
-    "vite": "^5.0.0"
+    "typescript": "^5.0.0"
   },
   "packageManager": "yarn@4.9.1+sha512.f95ce356460e05be48d66401c1ae64ef84d163dd689964962c6888a9810865e39097a5e9de748876c2e0bf89b232d583c33982773e9903ae7a76257270986538"
 }
