+++ dev/src/actions/predict-input.ts
@@ -1,0 +1,8 @@
+'use server'
+
+import { predictInputContent } from 'predictive-textarea'
+
+
+export async function predictInput(text: string) {
+  return predictInputContent(text)
+}
