+++ dev/src/app/page.tsx
@@ -1,0 +1,7 @@
+'use server'
+
+import { DevPlayground } from '../components/DevPlayground'
+
+export default async function Home() {
+  return <DevPlayground />
+}
