+++ dev/src/page.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import './styles/globals.css'
+import './styles/dev.css'
+import Home from './app/page'
+
+export default function Index() {
+  return (
+      <Home />
+  )
+}
